requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --no-index --output-file requirements.txt requirements.in
#
asn1crypto==0.24.0        # via cryptography
backcall==0.1.0           # via ipython
bleach==2.1.3             # via nbconvert
certifi==2018.1.18        # via requests
cffi==1.11.5              # via cryptography
chardet==3.0.4            # via requests
cryptography==2.2.2
cycler==0.10.0            # via matplotlib
decorator==4.2.1          # via ipython, traitlets
entrypoints==0.2.3        # via nbconvert
faker==0.8.12
html5lib==1.0.1           # via bleach
idna==2.6                 # via cryptography, requests
ipykernel==4.8.2          # via ipywidgets, jupyter, jupyter-console, notebook, qtconsole
ipython-genutils==0.2.0   # via nbformat, notebook, qtconsole, traitlets
ipython==6.3.1            # via ipykernel, ipywidgets, jupyter-console
ipywidgets==7.2.0         # via jupyter
jedi==0.11.1              # via ipython
jinja2==2.10
jsonschema==2.6.0         # via nbformat
jupyter-client==5.2.3     # via ipykernel, jupyter-console, notebook, qtconsole
jupyter-console==5.2.0    # via jupyter
jupyter-core==4.4.0       # via jupyter-client, nbconvert, nbformat, notebook, qtconsole
jupyter==1.0.0
kiwisolver==1.0.1         # via matplotlib
markupsafe==1.0           # via jinja2
matplotlib==2.2.2
mistune==0.8.3            # via nbconvert
nbconvert==5.3.1          # via jupyter, notebook
nbformat==4.4.0           # via ipywidgets, nbconvert, notebook
notebook==5.4.1           # via jupyter, widgetsnbextension
numpy==1.14.2             # via matplotlib, pandas, scipy
pandas==0.22.0
pandocfilters==1.4.2      # via nbconvert
parso==0.1.1              # via jedi
pexpect==4.4.0            # via ipython
pickleshare==0.7.4        # via ipython
pprintpp==0.3.0
prompt-toolkit==1.0.15    # via ipython, jupyter-console
ptyprocess==0.5.2         # via pexpect, terminado
pycparser==2.18           # via cffi
pycryptodome==3.5.1
pygments==2.2.0           # via ipython, jupyter-console, nbconvert, qtconsole
pyjwt==1.6.1
pyparsing==2.2.0          # via matplotlib
python-dateutil==2.7.2    # via faker, jupyter-client, matplotlib, pandas
pytz==2018.3              # via matplotlib, pandas
pyzmq==17.0.0             # via jupyter-client
qtconsole==4.3.1          # via jupyter
requests==2.18.4
scipy==1.0.1              # via seaborn
seaborn==0.8.1
send2trash==1.5.0         # via notebook
simplegeneric==0.8.1      # via ipython
six==1.11.0               # via bleach, cryptography, cycler, faker, html5lib, matplotlib, prompt-toolkit, python-dateutil, traitlets
terminado==0.8.1          # via notebook
testpath==0.3.1           # via nbconvert
text-unidecode==1.2       # via faker
tornado==5.0.1            # via ipykernel, jupyter-client, notebook, terminado
tqdm==4.20.0
traitlets==4.3.2          # via ipykernel, ipython, ipywidgets, jupyter-client, jupyter-core, nbconvert, nbformat, notebook, qtconsole
ujson==1.35
urllib3==1.22             # via requests
wcwidth==0.1.7            # via prompt-toolkit
webencodings==0.5.1       # via html5lib
widgetsnbextension==3.2.0  # via ipywidgets
